Added Player health regeneration, ground texture, and healing items
Player health doesn't regenerate until the player hits 30% health and only regens to 50% so that the healing items have a purpose, I also increased the players health to 20 and increased the zombie damage a little bit. There are three types of med kits which are just weak, normal, and strong med kits. I also tried to do a item slot thing for the healing items but it didn't work out.